+++ src/Nethermind/Nethermind.Evm/VirtualMachine.cs
@@ -1,4 +1,3 @@
-
 // SPDX-FileCopyrightText: 2022 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
@@ -20,7 +19,6 @@ using Nethermind.Logging;
 using Nethermind.State;
 
 using static Nethermind.Evm.EvmObjectFormat.EofValidator;
-using static Nethermind.Evm.VirtualMachine;
 
 #if DEBUG
 using Nethermind.Evm.Tracing.Debugger;
@@ -32,7 +30,7 @@ namespace Nethermind.Evm;
 using unsafe OpCode = delegate*<VirtualMachine, ref EvmStack, ref long, ref int, EvmExceptionType>;
 using Int256;
 
-public sealed unsafe class VirtualMachine(
+public sealed unsafe partial class VirtualMachine(
     IBlockhashProvider? blockHashProvider,
     ISpecProvider? specProvider,
     ILogManager? logManager) : IVirtualMachine
@@ -1363,81 +1361,4 @@ public sealed unsafe class VirtualMachine(
         _txTracer.ReportOperationRemainingGas(gasAvailable);
         _txTracer.ReportOperationError(evmExceptionType);
     }
-
-    internal readonly ref struct CallResult
-    {
-        public static CallResult InvalidSubroutineEntry => new(EvmExceptionType.InvalidSubroutineEntry);
-        public static CallResult InvalidSubroutineReturn => new(EvmExceptionType.InvalidSubroutineReturn);
-        public static CallResult OutOfGasException => new(EvmExceptionType.OutOfGas);
-        public static CallResult AccessViolationException => new(EvmExceptionType.AccessViolation);
-        public static CallResult InvalidJumpDestination => new(EvmExceptionType.InvalidJumpDestination);
-        public static CallResult InvalidInstructionException => new(EvmExceptionType.BadInstruction);
-        public static CallResult StaticCallViolationException => new(EvmExceptionType.StaticCallViolation);
-        public static CallResult StackOverflowException => new(EvmExceptionType.StackOverflow);
-        public static CallResult StackUnderflowException => new(EvmExceptionType.StackUnderflow);
-        public static CallResult InvalidCodeException => new(EvmExceptionType.InvalidCode);
-        public static CallResult InvalidAddressRange => new(EvmExceptionType.AddressOutOfRange);
-        public static CallResult Empty(int fromVersion) => new(container: null, output: default, precompileSuccess: null, fromVersion);
-
-        public CallResult(EvmState stateToExecute)
-        {
-            StateToExecute = stateToExecute;
-            Output = (null, Array.Empty<byte>());
-            PrecompileSuccess = null;
-            ShouldRevert = false;
-            ExceptionType = EvmExceptionType.None;
-        }
-
-        public CallResult(ReadOnlyMemory<byte> output, bool? precompileSuccess, int fromVersion, bool shouldRevert = false, EvmExceptionType exceptionType = EvmExceptionType.None)
-        {
-            StateToExecute = null;
-            Output = (null, output);
-            PrecompileSuccess = precompileSuccess;
-            ShouldRevert = shouldRevert;
-            ExceptionType = exceptionType;
-            FromVersion = fromVersion;
-        }
-
-        public CallResult(ICodeInfo? container, ReadOnlyMemory<byte> output, bool? precompileSuccess, int fromVersion, bool shouldRevert = false, EvmExceptionType exceptionType = EvmExceptionType.None)
-        {
-            StateToExecute = null;
-            Output = (container, output);
-            PrecompileSuccess = precompileSuccess;
-            ShouldRevert = shouldRevert;
-            ExceptionType = exceptionType;
-            FromVersion = fromVersion;
-        }
-
-        private CallResult(EvmExceptionType exceptionType)
-        {
-            StateToExecute = null;
-            Output = (null, StatusCode.FailureBytes);
-            PrecompileSuccess = null;
-            ShouldRevert = false;
-            ExceptionType = exceptionType;
-        }
-
-        public EvmState? StateToExecute { get; }
-        public (ICodeInfo Container, ReadOnlyMemory<byte> Bytes) Output { get; }
-        public EvmExceptionType ExceptionType { get; }
-        public bool ShouldRevert { get; }
-        public bool? PrecompileSuccess { get; }
-        public bool IsReturn => StateToExecute is null;
-        public bool IsException => ExceptionType != EvmExceptionType.None;
-        public int FromVersion { get; }
-    }
-}
-
-public interface IFlag
-{
-    virtual static bool IsActive { get; }
-}
-
-public struct OffFlag : IFlag
-{
-    public static bool IsActive => false;
-}
-public struct OnFlag : IFlag
-{
-    public static bool IsActive => true;
 }
